feat: OLED SPI驱动(厂商对齐)、双路电机(TB6612FNG)、双路霍尔编码器
- OLED: 完全对齐中景园1.3吋面板初始化序列(0xAD/0x8B DC-DC, 0xD9/0x1F预充电, 列地址0x02偏移)
- OLED: SPI初始化改由SysConfig自动生成，移除手动SPI_Init
- Motor: 双路电机A/PB13+PB15/PWMA/PA12  B/PB4+PB6/PWMB/PA21，统一MotorA/B API
- Encoder: AB相霍尔编码器双路(PB0+PB5, PB12+PB18)，GROUP1_IRQHandler四倍频计数
- MPU6050: yaw角互补滤波，备用magneto扩展接口
- Grayscale: 8路灰度传感器已就绪(PB19/17, PA16/14, PB20/25, PA25/27)
- 修复: GROUP1中断未使能NVIC，编码器和MPU6050 INT现已正常工作

+++ example/diansai/Hardware/mpu6050.h
@@ -41,8 +41,8 @@ void MPU6050_Calculate(void);   // 计算欧拉角（每5ms调用一次）
 void MPU6050_INT_Init(void);    // 配置 MPU6050 INT 引脚（数据就绪中断）
 extern volatile uint8_t MPU6050_DataReady;  // 数据就绪标志
 
-extern volatile float roll;     // 横滚角 (度)
-extern volatile float pitch;    // 俯仰角 (度)
-extern volatile float yaw;      // 偏航角 (度)
+extern volatile int16_t roll;     // 横滚角 (度)
+extern volatile int16_t pitch;    // 俯仰角 (度)
+extern volatile int16_t yaw;      // 偏航角 (度)
 
 #endif
